multi_modal_rag/data_processors/video_processor.py
# multi_modal_rag/data_processors/video_processor.py
import google.generativeai as genai
from PIL import Image
from typing import List, Dict

class VideoProcessor:
    """Process video content for key frames and visual analysis"""
    
    def __init__(self, gemini_api_key: str):
        genai.configure(api_key=gemini_api_key)
        self.model = genai.GenerativeModel('gemini-pro')
        
    def extract_key_frames(self, video_url: str, num_frames: int = 10) -> List[Image.Image]:
        """
        Extract key frames from video for analysis
        """
        # For YouTube videos, we'll use thumbnail and transcript
        # For actual video processing, you'd download and process the video
        
        frames = []
        # This is a simplified version - in production, you'd use cv2 to process video
        
        return frames
    
    def analyze_video_content(self, video_metadata: Dict) -> Dict:
        """
        Analyze video content using transcript and metadata
        """
        prompt = f"""
        Analyze this educational video:
        Title: {video_metadata['title']}
        Description: {video_metadata['description']}
        Transcript: {video_metadata['transcript'][:5000]}
        
        Extract:
        1. Main topics covered
        2. Key learning points
        3. Prerequisites needed
        4. Technical concepts explained
        5. Practical applications mentioned
        
        Format as structured JSON.
        """
        
        response = self.model.generate_content(prompt)
        return response.text