requirements.txt
# requirements.txt
opensearch-py==2.4.0
langchain==0.2.16
langchain-google-genai==1.0.10
google-generativeai==0.7.2
pypdf==3.17.0
PyMuPDF==1.23.8
sentence-transformers==2.7.0
youtube-transcript-api==0.6.1
pytube==15.0.0
pydub==0.25.1
arxiv==2.0.0
scholarly==1.7.11
feedparser==6.0.10
chromadb==0.4.20
gradio==4.8.0
python-dotenv==1.0.0
pandas==2.1.3
numpy>=1.25.2,<2.0
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2
whisper==1.1.10
yt-dlp==2024.3.10  # Alternative to pytube for YouTube data extraction